+++ spring-integration-core/src/main/java/org/springframework/integration/support/utils/IntegrationUtils.java
@@ -15,8 +15,13 @@
  */
 package org.springframework.integration.support.utils;
 
+import org.apache.commons.logging.Log;
+import org.apache.commons.logging.LogFactory;
+
 import org.springframework.beans.factory.BeanFactory;
 import org.springframework.core.convert.ConversionService;
+import org.springframework.integration.support.DefaultMessageBuilderFactory;
+import org.springframework.integration.support.MessageBuilderFactory;
 import org.springframework.util.Assert;
 
 /**
@@ -28,7 +33,17 @@ import org.springframework.util.Assert;
  */
 public class IntegrationUtils {
 
+	private static final Log logger = LogFactory.getLog(IntegrationUtils.class);
+
 	public static final String INTEGRATION_CONVERSION_SERVICE_BEAN_NAME = "integrationConversionService";
+
+	public static final String INTEGRATION_MESSAGE_BUILDER_FACTORY_BEAN_NAME = "messageBuilderFactory";
+
+
+	/**
+	 * Should be set to TRUE on CI plans and framework developer systems.
+	 */
+	public static final boolean fatalWhenNoBeanFactory = Boolean.valueOf(System.getenv("SI_FATAL_WHEN_NO_BEANFACTORY"));
 
 	/**
 	 * @param beanFactory BeanFactory for lookup, must not be null.
@@ -36,6 +51,43 @@ public class IntegrationUtils {
 	 */
 	public static ConversionService getConversionService(BeanFactory beanFactory) {
 		return getBeanOfType(beanFactory, INTEGRATION_CONVERSION_SERVICE_BEAN_NAME, ConversionService.class);
+	}
+
+	/**
+	 * Returns the context-wide `messageBuilderFactory` bean from the beanFactory,
+	 * or a {@link DefaultMessageBuilderFactory} if not found or the beanFactory is null.
+	 * @param beanFactory The bean factory.
+	 * @return The message builder factory.
+	 */
+	public static MessageBuilderFactory getMessageBuilderFactory(BeanFactory beanFactory) {
+		MessageBuilderFactory messageBuilderFactory = null;
+		if (beanFactory != null) {
+			try {
+				 messageBuilderFactory = beanFactory.getBean(
+						INTEGRATION_MESSAGE_BUILDER_FACTORY_BEAN_NAME, MessageBuilderFactory.class);
+			}
+			catch (Exception e) {
+				if (logger.isDebugEnabled()) {
+					logger.debug("No MessageBuilderFactory with name '"
+								+ INTEGRATION_MESSAGE_BUILDER_FACTORY_BEAN_NAME
+								+ "' found: " + e.getMessage()
+								+ ", using default.");
+				}
+			}
+		}
+		else {
+			if (logger.isDebugEnabled()) {
+				logger.debug("No 'beanFactory' supplied; cannot find MessageBuilderFactory"
+							+ ", using default.");
+			}
+			if (fatalWhenNoBeanFactory) {
+				throw new RuntimeException("All Message creators need a BeanFactory");
+			}
+		}
+		if (messageBuilderFactory == null) {
+			messageBuilderFactory = new DefaultMessageBuilderFactory();
+		}
+		return messageBuilderFactory;
 	}
 
 	private static <T> T getBeanOfType(BeanFactory beanFactory, String beanName, Class<T> type) {
